Introduce LogLine class
In upcoming commits, we're going to be processing a log file currently
only structured as a space-separated list with two values:

  * the path of the page that was requested
  * the IP address from which the request was made

Should we need to process logs with alternative structures or values in
the future, we're representing the lines with this class which we can
adjust in the future as needed.

As we're pressed for time, we're only introducing a simple validation
here to check that the correct number of values are present, but we
could perform more thorough checks here as needed.