Update data stored for a path when adding a new log line
In previous commits, we added the PathVisitData class to represent the
data for each path stored in our `@path_visits` dataset. This commit
uses that class to set the correct totals for any log line being added
with a path that already exists.

`.update` sees that the path already exists in `@path_visits` and uses
the return value of the block to determine what the correct values
should be. In this case, they are the incremented values provided by
the `PathVisitData` class. The `Set` being used for unique log entries
will return `nil` if the entry already exists, so we're using the
inverse of that to determine whether or not to increment the unique
visits count.