Introduce PathVisitDataReader
In recent commits, we introduced a data store class that would store a
hash of paths and visit data about those paths. As we want to be able
to produce custom output for that data store, this commit introduces a
'reader' class that will format that data as needed.

I'm running a little low on time here, so this isn't the most elegant,
but it builds a couple of arrays that we can output. Again, probably
not the most performant as we're having to store the whole data set,
but we're aiming to demonstrate that we can have a class that handles
the reading and outputting of the data separately from where it's
stored, and allows us to update formatting and output values as
needed.