[Illustrative commit] Add initial tests for desired output
The task asked to demonstrate an ability to use TDD, my approach to
this is often to start with a high-level
integration/feature/end-to-end spec of the desired output of the code.

I wouldn't normally commit this until the end of my sequence of
commits as it wouldn't be an atomic commit with passing tests, but
I've added these tests here as illustrative of the approach.

These aren't the exact tests I started out with, I've used the final
specs here to make the PR/commit more readable, but the spec was
essentially: using a dummy input file, check the output has the
correctly sorted and formatted output lines. Hope this is useful!