Introduce LogParser class
In previous commits, we introduced classes to store the data from our
log files, sort them and output them. We now want to introduce a class
that uses those classes to produce the desired output which can be
called from our script file.

Rather than adding new tests, we're un-commenting the tests we added
as an illustrative example in one of the initial commits for this
task. Now that the code has been implemented, the tests can now pass
and work within an atomic commit.