Update parser.rb to accept file path and call log parser
In a previous commit, we added the `LogParser` class that should
provide our desired output of sorted log entries for a given log file.

In this commit, we're adding a check for a valid file at the provided
path and calling the log parser. This should give us the required
output.