Introduce PathVisitData class
As part of our work to store data about each path referenced in a
given log file, we're introducing this class to represents the data
we'll store for each path. In upcoming commits we'll be introducing a
mechanism to store this data per-path.

'#increment_stats` provides us a mechanism to update the stats for a
given line in the logs. We always want to increment the visit count,
but we'll only want to increment the unique visit count if the visit
was unique.