Update PathVisitDatastore to store unique path and IP combinations
Within this class, we're already storing a list of visited paths and
the number of times each one has been visited (although those numbers
are not yet being incremented, that work is yet to come).

However we need to know whether a path has been visited by a specific
IP address before i.e. whether the visit is 'unique'. If the path has
been visited from this IP before, it's not a unique visit, if it
hasn't, it's a unique visit.

We're using a Set here which will only allow us to store unique values,
this saves us from needing to implement our own uniqueness logic.

As with the existing dataset, this is unlikely to be the most robust
storage solution for this list as we don't know the size of the
potential log files, however with the time constraints and the size of
the given dataset, this will work for our current requirements.

Alternatively, we could have stored this data in a database of some
sort, a text file etc. We could also look at the performance
implications of generating a hash of each unique log line to make the
lookup of an existing entry quicker.

This approach is currently also unique for each exact log entry, so if
URL params were added to the logs, they would currently be counted as
totally unique entries.