Introduce PathVisitDatastore class
As part of the work to produce a set of totals for the number of
requests a path has received within a log file, we need somewhere to
store these totals. This commit introduces PathVisitDatastore, which
is intended to act as a small class wrapping a hash where we'll be
storing the totals for each path.

In this initial commit, we're not looking to update/increment the
stored values, that will be added in an upcoming commit.

As the length of any other logfile that uses this script is unknown,
this storage approach is unlikely to be the most robust solution as
the entire thing will be stored in memory. However with the given
time constraints it'll be the most performant with the size of the
given dataset.

Given more time we could look to using a database, a text file, a
key-value store etc as alternative data stores and review the
performance implication of each approach.